Bsky_SAC_Finetuned/requirements.txt
# Enhanced SAC LoRA Training Requirements - Optimized for RTX 5070 Ti
# Core ML/AI packages with performance optimizations
torch>=2.1.0
transformers>=4.36.0
datasets>=2.14.0
peft>=0.7.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Performance optimizations (optional but recommended)
flash-attn>=2.3.0  # For faster attention computation
triton>=2.1.0      # Backend for flash-attn

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization and progress tracking  
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0

# Configuration and utilities
pyyaml>=6.0
pathlib
argparse
hashlib
logging
json
time
collections
gc

# Optional: for enhanced memory management
psutil>=5.9.0

# Optional: for better tokenization
tokenizers>=0.15.0

# Development tools (optional)
jupyterlab>=4.0.0
ipywidgets>=8.0.0
